organized stylesheets into styles folder
