+++ src/styles/App.css
@@ -32,7 +32,6 @@
 
 .topLevel {
   margin-bottom: 10px;
-  text-align: center;
 }
 
 #error {
