+++ src/styles/MutantDisplay.css
@@ -35,3 +35,12 @@
 .mutantName:hover {
     text-decoration: underline;
 }
+
+#save-file {
+    text-align: center;
+}
+
+#summary {
+    text-align: center;
+    margin:15px;
+}
